backend/ocr-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
easyocr==1.7.1
numpy==1.26.4
Pillow==10.4.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
